modbus master 并不全采用RTTHREAD 事件机制

+++ software/bsp/stm32f40x/drivers/24cxx.h
@@ -2,12 +2,12 @@
 #define __24CXX_H__
 
 
-#include "stm32f10x.h"
+#include "stm32f4Xx.h"
 
 
 
-uint8_t write_bytes(uint16_t addr,void *write_data,uint8_t num);
-uint8_t read_bytes(uint16_t addr,void *read_data,uint8_t num);
+uint8_t write_bytes(uint16_t addr,void *write_data,uint16_t num);
+uint8_t read_bytes(uint16_t addr,void *read_data,uint16_t num);
